requirements.txt
click
fastapi
folium
numpy
pandas
pint
pyserial
pyzmq
scipy
tabulate
utm
websockets
